+++ NEWS.md
@@ -1,4 +1,4 @@
-# kairos 2.2.1.9000
+# kairos 2.3.0
 ## Bugfixes & changes
 * Deprecate `simulate()`.